+++ src/interfaces.d.ts
@@ -1,0 +1,9 @@
+interface ToDo {
+	'file-path':string,
+	'file-line':number,
+	message:string
+}
+
+interface FolderStructure {
+	[folder: string]: ToDo[] | FolderStructure;
+}